+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## master (unreleased)
 
 * Added HEAD implementation (#7).
+* Document and make sure that the POST, PUT, DELETE and PATCH verbs would return a 405 error.
 
 ## v0.2.0 (2019-06-19)